tests/wpt/tests/css/css-ui/animation/caret-animation-parsing.html

<!doctype html>
<title>Parsing of caret-animation</title>
<link rel="help" href="https://drafts.csswg.org/css-ui/#caret-animation">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
<div id="target"></div>
<script>
  test_valid_value('caret-animation', 'initial');
  test_valid_value('caret-animation', 'inherit');
  test_valid_value('caret-animation', 'unset');
  test_valid_value('caret-animation', 'revert');
  test_valid_value('caret-animation', 'auto');
  test_valid_value('caret-animation', 'manual');
  test_invalid_value('caret-animation', 'none');
  test_invalid_value('caret-animation', 'manua');
  test_invalid_value('caret-animation', 'fade');
  test_invalid_value('caret-animation', 'blink');
</script>
